Income tags

+++ database/migrations/2020_01_01_000005_create_organizations_table.php
@@ -14,8 +14,8 @@ class CreateOrganizationsTable extends Migration
     public function up()
     {
         Schema::create('organizations', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('account_id')->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('account_id')->index();
             $table->string('name', 100);
             $table->string('email', 50)->nullable();
             $table->string('phone', 50)->nullable();
